Migrate fetcher tests to pytest
